requirements.txt
requests
PyJWT

tensorflow>=2.12
PyYAML
